Hide core word count display
It's not clear if we will keep the core word counter or use the one in TinyMCE. The core counter appears to only be accurate on page load (when v5 is loaded) and it is not auto-updated as typing occurs. This code change is not a final fix; it's just a cosmetic repair to get the core counter out of the way. On the opposite side of the editor (at lower-right) the accurate word count is provide by TinyMCE v5 and, I might add, you can click it to toggle between the word count and the character count.