Update Tiny config object
This change pares down some of the unneeded options that were included in the Tiny config object. It also adds a few new properties for the toolbar, editor height, and spellcheck. It also attempts to define a logical first and second row of buttons (presuming the final release will contain 2 rows of buttons.) The third row of buttons is "everything else".